qa/1844: needs a relatively recent remote pmcd
There was on version specified for the remote host, we need at least
PCP 4 (I think) ... in the QA Farm when this test is run on vm03,
it was picking vm35 as the remote host, only problem is vm35 is
running PCP 2.7.8 (for backwards compatibility testing) and the
required pmcd.pid metric was not there in PCP 2.7.8 (it was added
circa Sep 2017).

Setting the version required to >= PCP 4 (circa Feb 2018) seems safe
enough.