src/libpcp: assorted hardening (covscan)
We have multiple routines that are wrappers for read(), write(),
select(), etc. and Coverity is complaining that from the caller
context it is possible that the fd is negative ... harden these
routines that take an fd as an argument and return -EBADF if
it is not >= 0.

Hopefully this will clean up a bunch of Coverity CIDs in the caller
code.