qa: add _check_redis_server to common.check, update tests to use it
Add _check_redis_server to check the local redis server is
up and running and the version is not too old. The existing
_check_series function now only checks if redis is installed
(but not whether the server is running, as is the case in CI).
Note _check_redis_server also calls _check_series so it's
never necessary to call both in a test.

Of the 33 tests that were previously _notrun if _check_series
fails, only 1212, 1457, 1661, 1696 and 1824 needed to be changed
to _notrun if _check_redis_server fails. The rest are unchanged
and still use _check_series (so should now be run in CI).

There seems to be a new valgrind leak in 1695 if the local redis
server isn't running - this is still under investigation but seems
to be due to error handling when redis_connect fails (so this
test may end up being changed to use _check_redis_server).