qa/1695: update to use it's own redis server, add repeated scrapes
Updated this problem child to start it's own redis server, so it'll
run in CI (after checking redis is new enough). Also added a new
subtest to run repeated /metrics scrapes of only one metric using
the default poll-timeout. This leaks with the following back-trace
sometimes - it would be interesting to see if any platforms do not
leak with the backtrace (and also whether any platforms always do).

> 28 (24 direct, 4 indirect) bytes in 1 blocks are definitely lost in loss record 44 of 283
> at 0x4839809: malloc (vg_replace_malloc.c:307)
> by 0x4C386FB: zmalloc (util.c:73)
> by 0x4C443C2: dictAddRaw (dict.c:304)
> by 0x4C442C3: dictAdd (dict.c:261)
> by 0x4C52F2C: webgroup_new_context (webgroup.c:245)
> by 0x4C5367E: webgroup_lookup_context (webgroup.c:369)
> by 0x4C583FE: pmWebGroupScrape (webgroup.c:1907)
> by 0x11F797: pmwebapi_scrape (webapi.c:923)
> by 0x4C8C7AD: worker (threadpool.c:122)
> by 0x4F713F8: start_thread (pthread_create.c:463)
> by 0x4E9EB52: clone (clone.S:95)