dbpmda: add -v flag for valgrind
When debugging a PMDA, it is useful to be able to run dbpmda under
the control of valgrind to exercise a DSO PMDA (so you can see,
for example, memory leaks in the PMDA).

The -v flag stops dbpmda from dlclose()'ing the DSO PMDA on exit,
which allows valgrind access to the DSO PMDA's symbol table and
makes the valgrind report a whole lot more useful.