pmFreeLabelSets(3): small change in semantics
The code paths available for building and returning pmLabelSets are
really quite complex and involve any of these routine (some of which
are nested):
    __pmAddLabels, pmGetClusterLabels, __pmGetContextLabels,
    pmGetContextLabels, __pmGetContextLabelSet, __pmGetDomainLabels,
    pmGetDomainLabels, pmGetInDomLabels, pmGetInstancesLabels,
    pmGetItemLabels, pmLookupLabels, pmMergeLabels, pmMergeLabelSets,
    __pmParseLabelSet, __pmParseLabelSet.

As errors can occur at any level, the resulting pmLabelSet structure
may be missing entirely (the pointer is NULL) or incomplete ... to
reduce memory leaks and Coverity issues, it is necessary to call
pmFreeLabelSets() even on the error paths after calls to any of the
routines above.

The code will be a whole lot simpler if pmFreeLabelSets() simply
handles the NULL pointer case as a no-op.

This will also avoid potential SEGV issues in pmFreeLabelSets()
where sets[i] was used, even if sets == NULL.