qa: cull proc.psinfo.tty_pgrp from qa/1222 output
This metric might be returned from the kernel with a value
of -1 - the PMDA drops all such instances (processes) from
the indom for this metric, and under container QA (ie with
very few processes) we can end up with <10 instances where
there are few/no ttys.  That ends up causing indeterminism
in the test, so cull just that one metric.