BCC PMDA: make netproc the only default module
The recently added netproc module is more accurate and complete than
the earlier tcpperpid so it makes a much better default module.

There was no clear reason for others being defaults so make netproc
the only default module to minimize the impact of the default PMDA conf.

Related to issue #1234.