pmrep et al: allow specifying instances with PID
Similar to pmval(1), allow specifying instances with PID (as part
of the instance name for proc metrics) without the need to provide
full instance name string or regex:

$ pmrep -C -i 1,11,111 proc.psinfo.pid
  p.p.pid  p.p.pid  p.p.pid
  000001   000011   000111

Full QA passes so no regressions. If no surprises during a release
or two this change will be documented to make this change official.

Related to issue #1236.