qa/886: major overhaul
Poor old pminfo caches the indom (good) but then uses an O(N*N)
algorithm (bad) to lookup instance id -> instance name for the
N instances in the indom ... given a big enough N (like for the
interrupt lines per CPU on an 1152 CPU machine) this will kill you.

So use qa/src/indom -r 10 -v to sample instances at 0%, 10%, ..., 90%
and 100% through the indom instead of pminfo.  Has close to the same
testing coverage, at a fraction of the cost.

Reduces output from 53Mbytes down to under 18Kbytes ... so the bzip2
compression of 886.out is gone.

Running time (after changes to linux PMDA interrupts indoms) is down
for 4600sec to 49sec on bozo.