src/libpcp_pmda/src/cache.c: don't cripple hash tables
For some reason, the number of hash chains per hash table was
capped at 1024 ... if you have a really big instance domain
(like in qa/886 where this was noticed), then each hash chain
can be very long with a consequent performance hit.

Keep the number of hash chains to about 1/4 the number of entries
in the hash table, but drop the 1024 cap.