qa/tmparch: stiffening to protect pmlogger from PMDA autoinstall
Just after a package install/upgrade, we may have a bunch of PMDAs
that need to be automatically installed or removed the next time
pmcd is restarted.  If pmlogger starts running to create QA archives
while this is happening, we risk getting unwanted <mark> records in
the QA archives.

Check and if necessary wait a while until all clear ... done in
the new check-pmcd-stable script.