ci: rename "Daily QA" to "QA", always generate Allure report
When testing larger commits, it's advised to run a full QA run on the
new branch before merging. This change generates the Allure test report
when the QA workflow is trigged manually, and publishes the report as a
zip file.

I've renamed "Daily QA" to "QA", as it can and should be invoked manually
as well. When the workflow is run on schedule, it publishes the test
report to GitHub pages and sends a Slack message as usual.