pmconfig.py: more consistent error handling
pmconfig.py methods use IOError exception (as it should) for basic
config file access related issues when e.g. opening a file fails.

In some cases when the contents are not valid or metrics are missing
either IOError is raised or an error is printed before exiting. So
things are not completely consistent at the moment.

IOError doesn't seem to be completely fitting for configuration
errors. OTOH, hard exit from general helper methods in theory
might not always be the best choice if a client tool would try
to recover from those issue. But given that no in-tree tools
try to recover and without any valid metrics it's a bit hard to
see what else a tool could do at that point than exit.

This patch makes error handling a bit more consistent by keeping
IOError for low-level file operation related errors but otherwise
always printing out an error message everywhere before exiting.

In case a need arises in the future to raise an exception then
that could be considered but that should happen then consistently
(and probably using something else that IOError), not just in one
or two locations.

No currently known functional changes by this patch.