qa/1315: check for absence of <mark> records in tmparch archives
These are created as a preliminary step in QA, and for the most part
should not contain <mark> records.

A recent problem in CI was first diagnosed (after some considerable
difficulty) as <mark> records in tmparch/foo, so this test is an
attempt to check for that, and it is in the sanity group so it gets
run as part of the normal CI QA and hopefully will provide early
warning if the problem reappears.