qa/new-grind: new helper script
Given an exsting test created with a recent qa/new (i.e. includes
the valgrind hooks), create a new test that runs the existing test
with --valgrind to exercise valgrind/helgrind on the same test case(s).

The new test is effectively a wrapper for the old test, but new-grind
does all the associated group changes, comment editing and git
leg-work.