Man page section sorting update
Section in manual pages we not in any particular order, pick the
one used in PCPIntro(1) and reorder the rest according to it.

The order is now as follows (when sections exist):

OPTIONS
EXAMPLES
NOTES
CAVEATS
DIAGNOSTICS
FILES
ENVIRONMENT
PCP ENVIRONMENT
AUTHORS
SEE ALSO

While at it, drop some DIAGNOSTICS section which were stating
the obvious (diagnostic messages intended to be self-explanatory).