libpcp_web: handle alloc failure in redisSlotsRequest
In redisSlotsRequest(), if redisSlotsReplyDataAlloc()
fails and returns NULL the async redis request still goes
ahead but with a NULL private data pointer passed down
the call stack ... leading to badness (much later on)
when the reply data is eventually dereferenced.

So rather than proceed with the request, return -ENOMEM
to the caller. There is already a related ENOMEM return
in redisSlotsRequest(), but none of the callers check
for it.

It is unacceptable to exit due to ENOMEM in the library call,
so that is not an option. There are over 50 redisSlotsRequest()
calls that will hence need error handling (future work).

Made similar changes to redisSlotsRequestFirstNode().

Also fixed up a C99 style declaration, which is not portable
on older platforms.