selinux: use new logging_watch_all_log_dirs_path macro
Allow PMDAs that access log files to do so when latest versions
of the kernel and selinux-policy is installed.

For some strange reason we've had to switch to the policy_module
macro instead of using a 'module' declaration directly - without
this we see compilation errors with this change.  See Red Hat BZ
https://bugzilla.redhat.com/show_bug.cgi?id=1966397

This resolves a new qa/652 (pmdasystemd) failure amongst others.