qa: check if the RediSearch module is installed, not if it's loaded
qa/1871 and qa/1872 start their own Redis instance with the module
loaded, so it doesn't matter if the RediSearch module is enabled for the
local running Redis server (if any is running).

This way the tests also run in CI, where the default Redis configuration
(which doesn't enable RediSearch) is used.