docs: drop commented env vars from command line option configs
The pmcd and pmproxy options files have not supported setting
environment variables via their .options file for many years.
These files are for command line options only and the scripts
enforce that, dropping all lines that do not start with minus
character.

Environment variables are only set via the sysconfig files -
make some improvements to the documentation so this is clear.

Resolves Red Hat BZ #1967174