qa/common.filter: refine _inst_value_filter()
When dealing with proc.psinfo.environ in qa/022, we sometimes run
into this ...
    proc.psinfo.environ
	inst [1 or "000001 /usr/lib/systemd/systemd"] value ""
	inst [38560 or "038560 /bin/sh"] value "... BASH_FUNC_which%%=() {  ( alias;
     eval ${which_declare} ) | /usr/bin/which ... --show-dot "$@"
    } _=./022"
	inst [1122 or "001122 /usr/bin/abrt-dump-journal-xorg"] value ""

and the old filter was looking for " at the end of the line to
terminate the string that starts at value " ... unfortunatley this
quits at --show-dot "$@", leaving the extra line
    } _=./022"
in the output.

The new filter counts " and continues until a " is found at the end of
the line and the " count is even (so it balances the " that starts at
value ") ... this is not bulletproof (backslash escaping or embedded
quoting will still confuse it) if that happens I'll have to adopt a
more draconian regime.

Seen on vm26 (Fedora 34) where the value of proc.psinfo.environ for
a /bin/sh process is some 46 Kbytes long in this test case.