qa/917: rework for sedismod(1) changes on Fedora 34
As of Fedora 34, it appears that the output from sedismod(1) has sorted
the attributes in the rules.  This breaks any chance to compare them
directly with the original policy rules.

Use a new _sort_attr() function in qa/917 to sort attributes within a
rule and apply this to both the expected output (from qa/917.out.in)
and the output from sedismod(1).

Clean up some remnants of the rpm PMDA while I'm here.