pmcd pmda: add pmcd.zoneinfo
This is intended to be used in the V3 archive format where we're
planning to move away from using local timezones in the style of
"AEST-10" or "EST-11EST-10,87/2:00,297/2:00" and adopt something like
":Australia/Melbourne".

pmcd.zoneinfo uses heuristics to find the current system's tzfile
starting from /etc/localtime and if necessary going on an archaelogical
dig below /usr/share/zoneinfo ... this is not as simple as one
might hope as the new code in getzoneinfo() and getzoneinfo_plan_b()
demonstrate.  However the code works on Linux and BSD systems when
/etc/localtime is either a symbolic link or a copy of an installed
tzfile.

qa/1328 exercises this.