qa/1329: (new) exercise various code paths for help text
Especially cases where
(a) there is no full text, just the one line, and
(b) the help text does not come from newhelp(1) generated files

The pmproxy PMDA fills the bill!

And was failing qa/408 ... this test is a more strenuous check
of the "fallback" code paths in libpcp than qa/408.