src/libpcp/src/help.c: Fix "fallback" logic
Since Day One, the rule has been that when asking for "full" help
text, if none is available then return the "one line" help text if
that is available (else PM_ERR_TEXT).

The tricky part is that this could be implemented in one or more
of:
(a) the PMDA,
(b) libpcp_pmda (if the PMDA uses it)
(c) libpcp, or
(d) the application (pminfo or pmchart, for example).

When chasing a bug with the help text for the pmproxy PMDA (that
turned out to be a totally unrelated issue ... a stale and old format
mmv file), I noticed that the code for (c) looked wrong.

To make matters worse the code depends on the context type as
PM_CONTEXT_HOST, PM_CONTEXT_ARCHIVE and PM_CONTEXT_LOCAL all have
different protocols for reporting "full help text is not available"
(assuming (a) and (b) do not apply).

This commit tries to make the logic in lookuptext() and in particular
the conditions under which fallbacktext() is called more correct.