qa/134: tweak range arithmetic
The test reports a (non-deterministic) change in archive size by
doing some truncated rounding of the archive size.

Unfortunately the size includes the logging of
the hostname, and in CI this can be something like
fv-az121-446.xhy4bhhdincepo0av04agjzkjc.cx.internal.cloudapp.net
whose length exceeded the original assumptions in the arithmetic.