pmdahacluster: Remove un-needed pmdaCacheLookupName() calls
Update instance refresh code for the label supported metrics to remove the
superfluous pmdaCacheLookupName() calls.

We update the label supported indoms in mirror to the non-label indoms
and in turn don't require re-checking instance updates immediately again
(after doing so for the non-label indoms).

This also fixes the Coverity defects reported by Nathan for unchecked returns
on these calls.