pcp-ss: fix EAGAIN exceptions, add filter header, QA
Use explicit scale="instant" in fetchgroups to avoid rate conversion.
This was causing fetchgroups for metrics with counter semantics
to throw EAGAIN exceptions in the report function. pcp-ss reports
the raw counters, but other tools such as pmrep want to do rate
conversion, so we keep the correct (native) semantics.

Also simplify the code a bit, add missing bytes_sent in the --info
report. Add a header in the report for the current timestamp and
filter, which is especially useful for archive replay since the
filter can change over time whilst logging (it's a storable metric).

Tweaked pcp-ss report in qa/1634.out