qa/600 & qa/src/tztest.c: finally tired of this failing twice a year
This test historically breaks in the first week of July and January
EVERY year.  It included this comment:
    I don't understand the sed replacement below - Ken, Jul 2003
and 8 years on nothing has changed.

So I've refactored the test program and removed the silly 1st
half of the year vs 2nd half of the year logic, set the test
datestamps correctly and nuked the undecipherable sed.

Let's see what happens now ...