libpcp: redefine __pmLogSetTime()
The problem is that if physical files for the data volumes of an
archive are removed (asynchronously by someone else) while we're
trying to switch volumes then we don't handle this safely.

The previous commit 10763319 was as stop-gap to address Mark's SEGV
issue at https://github.com/performancecopilot/pcp/issues/1338 and
simply handled direct calls to __pmLogChangeVol() and ensured the
return status was checked.

I was aware, then Coverity made a lot more people aware, that this
"fix" was incomplete, specifically the calls to __pmLogChangeVol()
from within __pmLogSetTime() were not checked.

To fix the latter we have to change the type of __pmLogSetTime() from
void to int so we can return status to indicate that __pmLogChangeVol()
has failed.  And then make sure all the callers of __pmLogSetTime()
check the status returned from that function.

    modified:   src/libpcp/src/fetch.c
    modified:   src/libpcp/src/internal.h
    modified:   src/libpcp/src/logutil.c

Because this introduces some new -Dlog diagnostics, qa/251 needed
a bit of a make-over.