qa/1600: number of values is non-deterministic
The number of
    [TIMESTAMP] VALUE
lines in the "verify metric values" test is non-deterministic (and
depends in asynchronous activity between pmlogger, pmproxy and a
Redis server ... for which pmsleep just does not cut the mustard).

We are expecting 2 lines, but often see only 1, especially in CI.

So, filter the output and accept either 1 or 2 lines of "values".