src/pmlogger/src/fetch.c: tighten up "lost pmcd connection" logic
We're seeing a failure of qa/964 (regularly in CI, less often
elsewhere) where the PID of the primary pmlogger changes
unexpectedly.

By slavaging the pmlogger.log.prev in that QA test, I can see
    myFetch: End of File: PMCD exited?
    pmlogger: Error: IPC protocol failure
    pmlogger: Lost connection to PMCD on "8024ad50d844" at Sun Jul  4 21:05:13 2021
    pmlogger: pdu.c:139: pduread: Assertion `fd >= 0' failed.
which is all as expected EXCEPT the last line ... once the connection
to pmcd has been marked as down, we should not be trying to
send/receive PDUs until a pmcd reconnection is made.

In the heart of myFetch() it looks like we may have received a
pmcd state change just before the connection was lost, and then
continue to try and process that state change, which is wrong.