qa/189: dodge pmie sleepTight() babble
For slow VMs, like bozo-vm, we sometimes see .out polluted with diagnostics
like:
    sleepTight: negative delay (-1.680389). sched=1625522929.467434, cur=1625522931.147823
    Last sleepTight until: Tue Jul  6 08:08:49.367434 2021
    Last Task dump @ 0x55eb658a7ff0
      nth=0 delta=0.100 tick=76 next=0x55eb658a71c0 prev=(nil)
      eval time: Tue Jul  6 08:08:49.467433 2021
      retry delta: N/A host: bozo-vm.localdomain
      via=local: (up)
      rules:
	expr_1

Since this test is using 100msec sampling rate, and if the VM
misses the host scheduler boat the delay can be O(1sec), meaning
all our fetches are late.  But this is irrelevant to this test which
is trying to see if the children processes of pmie are waited for
and terminated correctly.

So send skip pmie's stderr (other than for the pmie exiting message)
in the .out file (it was already being appended to the .full file).