src/pmlogger/src/pmlogger.c: block SIGALRM before re-exec
Current theory is that qa/793 failure (pmlogger sometimes vanishing)
may be due to the kernel delivering a SIGALRM to pmlogger that was
set just before the re-exec and delivered after the re-exec but before
the signal handlers are installed.

Currently we don't do any signal fiddling until after the configuration
file has been parsed ... which matches the qa/793 apparent failure
area in the code.