qa/common: don't dink with PCPQA_SYSTEMD
When stopping/starting/restarting pmcd and/or pmlogger in havagents(),
havelogger() and havepmcdtrace(), use _service without unsetting
$PCPQA_SYSTEMD.

I don't recall why we were avoiding systemd here, but it is a really
bad idea to not use systemd when systemd is _really_ in control of
these processes.

We had got away with it until commit b16482ed6 changed a bunch of
lines in common.check from
    if `which systemctl >/dev/null 2>&1`
to
    if [ "$PCPQA_SYSTEMD" = yes ]
on the assumption that we don't need to keep checking if systemctl
is available once PCPQA_SYSTEMD is set in common.check.

And this led to a meltdown in CI.