libpcp/pmapi.h/pmdbg: add -Ddev0,dev1,dev2
These are "developer" only flags ... feel free to add/remove diags to
support this week's development efforts, and then rip 'em out or move
them under one of the other -D flags if they are intended to persist.

Except for long-running development tasks, there should be no
references to pmDebugOptions.dev0 (or dev1 or dev2) in the code base
at the time of a release.