qa/1573: add diags for pmproxy maps
cat /proc/PID/maps to 1573.full. This shows libcap.so.2
is linked with pmproxy on f34 but not f33. That could explain
some or all of the larger startup RSS on f34. Doesn't explain
why pmproxy needs ~ 10MB RSS on startup (on any platform),
but at least we have a plausible explanation for the f34 RSS
difference.