src/pmlogger/src/callback.c: fix "emit duplicate indom records" bug
This has been here since before Sep 3 2002 (the oldest version of
callback.c I still have access to), and the logic is wrong.

The heuristic used to decide if an indom record should be emitted into
the .meta file is too weak, and under special (but rare) circumstances
it can lead to the identical indom appearing multiple times in the
.meta file.

Other than a bigger .meta file and a larger memory footprint when
the archive is opened for reading, there are no ill-effects from
this, and the change should be transparent to all users.