sample PMDA: add sample.proc.* metrics
These metrics are for "fake" processes that mimic the proc PMDA,
namely a dynamic instance domain with some long-term instances, some
short-term instances that randomly appear and disappear, monotonic
increasing "pids" that eventually wrap, but are guaranteed to be
unique.

The difference is that these metrics and the underlying instance
domain is deterministic and can be reset (via pmstore) to a known
starting point.

These metrics will be useful for the V3 archive work, and in particular
the "delta indom" changes.