libpcp: initial data structures, macros and TI for v3 archives
Adds important macros and data structures we anticipate needing
for version 3 of the PCP archive format.  Code to read, write,
and make use of the new temporal index format is included here.

There should effectively be no change (yet) from this commit as
new code is added under the __PCP_EXPERIMENTAL_ARCHIVE_VERSION3
macro (which is unset) for the most part.