__pmResult: introduce typedef and some helper routines
This will (later) become part of the Y2038-safe work, when all internal
uses of pmResult will be promoted to use __pmResult with more bits
for the seconds field and nanosec precision.

Also included is a new src/hacks/pmresult-fiddle which is a proof of
concept for the pmResult <--> __pmResult transition.