pmdaproc: ensure ENODATA error code not exposed on fetch callbacks
Ken encountered a situation where the proc.id.container metric, as
well as proc.id.label and others, was causing cryptic messages to
be logged on fetch.  This is because libpcp_pmda expects specific
error codes only to represent a lack of values - code now ensures
only PM_ERR_VALUE is presented for this case.