__pmLogLabel: changes to support log version independence
Internally, both v2 and v3 labels features are accessible
via the (common) __pmLogLabel structure.  A __pmTimestamp
is used therein for all log start time references.  We're
now using dynamically allocated strings for the hostname,
timezone and zoneinfo fields as well.

A stub __pmZoneinfo routine is added to libpcp - this will
be filled out with the pmcd.zoneinfo metric code via a new
commit, and pmdapmcd will share the implementation as it's
going to be used for local contexts and v3 archives also.

pmNewContextZone now uses the pmcd.zoneinfo metric when it
is available.

Next round of on-disk support for v3 archives is included
here too, under __PCP_EXPERIMENTAL_ARCHIVE_VERSION3 still.
Finally I've dropped qa/pmlogconv as this function will be
handled by pmlogrewrite for v2 to v3 conversions.