pmlogextract, pmlogrewrite: move away from "PDU" references
These tools read and write PCP archives, they do not communicate with
pmcd (or anything else) using sockets and PDUs.

The on-disk structures for PCP archives are, for the most part, made
up of records that are formatted into multiples of 32-bit values, so
it makes more sense in the code for these to be arrays of __int32_t
rather than __pmPDU.  Also recbuf is a more consistent name for a
variable than pdubuf after the type change.

Another reason is that __pmPDU is unsigned, and the recent
__pmTimestamp work has thown up some corner cases where we need
to convert a 32-bit value on-disk to a 64-bit value in memory (and
vice versa), and this may produce unexpected results if the value is
negative unless the source and destination are treated as signed.

There is some overlap of data types, because some of the on-disk
records are in fact copies of PDUs from pmcd, e.g. a pmResult, so we
still need some (__pmPDU *) casting love in a couple of places.

This commit makes the source for these tools more consistent with
the changes that have already been made in libpcp.