pmdapodman: correct the metric count calculation for pmdaInit
This was observed on a RHEL test machine to produce failure to
launch the PMDA, with an odd logged message relating to metric
indom serial number checks (as it accesses a few bytes beyond
the end of the metric table).  Existing QA tests are been able
to detect this failure mode.

Related to Red Hat BZ #1962019