services: switch logutil and pmieutil scripts from type oneshot to exec
Switches the "daily" and "check" and related services from Type=oneshot
to Type=exec (or Type=simple for older versions of systemd that do
not support Type=exec) and nuke KillMode=none. Explicitly specify
Restart=none, even though it's the default, to be clear that these
timer invoked service scripts are not long running service daemons
and are expected to exit normally.

These changes are a precursor to further improvements to the long
running PCP services and systemd integration.

All tests in the logutil and pmieutil QA groups are passing and soak
testing on several platforms has not shown any new issues (and the
systemd warnings about "Support for KillMode=none is deprecated" are
gone).

Resolves: RHBZ#1942844
Resolves: Fedora BZ#1897945
Resolves: https://github.com/performancecopilot/pcp/issues/1186
Obsoletes the earlier PR#1355.