pmieconf: improve the CPU saturation rules
Resolves some formatting issues with the many_util pmie rule.
The back-quoting was incorrect, resulting in quotes appearing
in the syslog message strings.  The message accidentally used
the %c (source) pmie formatting string in %cpus as well.  One
other change here is to refactor the many_util rule so that a
singelton metric is the first part of the expression, not the
per-CPU metrics, so that the %h pmie format expansion gives a
single string when the rule evaluates to true.

Finally, bump up the min/max CPU count threshold that we use
to indicate a small/moderate-sized server (from 4 to 12) - we
use this to transition between different rules' effeciveness.