ci: refactorings
Previously the platform.yml files were structured around the idea that
the same set of commands can run independent of the environment (direct,
container, virtual machine). It turns out that this wasn't feasible
always, and limiting in other cases. To make room for special use cases
- for example a configuration with Ubuntu 20.04, direct (no container),
and additionally BTF enabled - I've removed this abstraction and every
platform.yml is now of a specific type (direct or container).

The VM runner/type is deprecated, as this didn't work well in
GitHub Actions due to missing nested virtualization support of the
build machines.

Other changes:
* replaced the 'install_build_dependencies' with a generic 'setup' step
* working directory of all tasks is now the pcp repository, artifact dir
  is ../artifacts
* dropped network issues workaround (should be stable now afaics)
* code formatting (`black --line-length 120`)