pmieconf: update perdisk rule help text
Turns out that the pmieconf rules per_disk.bandwidth and per_disk.iops
only work (out of the box) if you have a disk named "sda".

The help text to describe how to reconfigure this to a real (set)
of disk names was a bit misleading.

qa/1049 was clueless about this, and duly failed on any machine that
did not have an "sda" disk ... of course the QA Farm had one machine
like this.

Another example of a long-standing QA failure in a corner case that
has only become apparent as the number of QA failures is down to
single digits (across 10's of thousands of test cases)