src/pmlogextract/pmlogextract.c: fix endian botch for timestamp in -w <mark> record
Not sure how this escaped previously, but qa/270 started failing
after recent __pmTimestamp changes ... root cause seems unrelated
to the __pmTimestmap changes, but who knows?

The <mark> record written at the end of a -w window did not have
the correct endian conversion for the timestamp.  The way the
<mark> record is used it is probable that no-one has ever (before)
looked at the timestamp in this record.