libpcp_web: avoid duplicate instname labels in /metrics response
If a PMDA already returns instname and instid labels for a
non-singular metric, don't insert them again in the /metrics
webscrape response because they will be duplicates. Duplicate
labels are not OpenMetrics compliant and break some downstream
consumers of the /metrics end-point, e.g. prometheus.

qa/1191 and 1221 are remade with duplicate_hostname_label.txt,
which tickles this bug. An additional QA test to exercise the
updated webapi /metrics code is still coming.