pmlogger: add -I/--pmlc-ipc-version command line option
Due to a bug in libpcp, autonegotiation of the IPC version between
pmlc and pmlogger was broken for a down-rev pmlc and an up-rev pmlogger.

The new command line option clamps the highest IPC version number that
pmlogger will offer to pmlc, providing a workaround for the libpcp
bug.

Also a couple of -Wshadow warnings fixed and change from -Dcontext
to -Dpmlc for the diags on the pmlc-pmlogger channel.