pmproxy: more robust fix for instname/instid duplicates
Reworks the previous change to ensure duplicate labels for
openmetrics scraping are not created, for the special case
of the instname and instid pseudo-labels.

This removes the use of strstr to detect instname inside the
JSON label string as this approach produces false positives
when instname= is embedded within a label value or simply a
similar label name like myinstname.  The approach now makes
use of the pmLabelSet information before merging it to form
the final combined labels string.

Related to #1407