configure.ac: -g is sometimes a bad choice for gcc
Specifically on OpenBSD 6.8 and the i386 architecture, if gcc is
fed big source files (and we have a bunch of 'em), then it may emit
debuginfo code with "offsets" that are illegal.

The change here is very targeted, and there is no change for distros
other than OpenBSD, other versions of OpenBSD or amd64 builds for
OpenBSD 6.8.