libpcp: improve memory footprint of pmSetProcessIdentity
initgroups(3) has been observed to be expensive in terms of
its RSS memory footprint - which blows out the PCP daemons
(all call this, including PMDAs) - as a result getgroups is
used to check first and call initgroups only when needed.

Usually its not as the PCP_USER has no supplementary groups
by default.