libpcp_fault: let me play in the big kids' pool
Over 10 years ago a fault injection subsystem was developed for libpcp.

In the intervening period it has languished, with only the odd
bozo-developer even trying to use it in manual QA.  This is unfortunate
because libpcp_fault and the associated QA group "fault" exercise
some rarely trodden code paths.

The recent __pmResult (aka __pmAllocResult()) changes have re-activated
interest in fault-injection testing.

This commit:
- forces the build to make libpcp_fault
- packages the libpcp_fault libraries and associated <pcp/fault.h>
  header into the pcp-testsuite package (both dpkg and rpm)
- removes qa/changeversion that is no longer relevant and clobbered
  libpcp_fault