pcp-atop: handle different event naming of ix86arch PMUs
The ix86arch PMU, and hence [ix86arch] section of perfevent.conf,
uses a different (singular form) name for INSTRUCTIONS_RETIRED to
everyone else.  Deal with this quirk such that we fallback to the
alternate name if the regular plural form cannot be found.

New test qa/1978 emulates this scenario using pmlogrewrite.

Resolves Red Hat BZ #1986264