src/pmlogger/src/fetch.c: fix small error in EOF logic on pmcd channel
pmlogger is supposed to hang around when pmcd exits, and when
pmcd restarts then pmlogger checks if it is able to keep going.

There was a small error in the logic, and if pmcd exited while
pmlogger was trying to do a pmFetch (or the pmlogger version,
myFetch), then pmlogger would exit on a subsequent __pmGetPDU
with a bad fd ('cause the socket has been closed by the EOF).

This manifested as qa/964 failing sometimes, but always in a way
that seemed impossible to reproduce ... because of pmcd exists
when pmlogger is NOT doing a pmFetch (i.e. most of the time) the
problem does not exist.

Fix this and add some additional diags on the "pass" code path
in qa/964.