libpcp: add PM_ERR_BOTCH
I've encountered an error condition (in the Linux PMDA) where we've
made assuptions that are incorrect and as a result a /proc statsfile
parser silently does the wrong thing.

But in fixing these I realize we don't have an error code that matches
the desired semantics, so
    PM_ERR_BOTCH (Internal inconsistency detected or assertion failed)
seems appropriate.

Fixing the problem in the Linux PMDA is the subject of another commit.