libpcp, pmcd: initial code for a new pmLookupDescs PMAPI routine
This implements handling of a new PDU in libpcp that allows bulk
processing of metric descriptors.  The aim is to assist with the
reduction of round-trip PDUs during PMAPI client startup, which
has been observed to be expensive over high latency networks for
tools dealing with many hundreds of metrics (like pmlogger).

New pmcd code is added and the PCP protocol is extended to allow
new clients to detect remote pmcd support is also in place.  The
implementation for archives and local context mode simply uses a
loop around pmLookupDesc, effectively, and special case handling
is in place for derived metrics.