pcp-atop: report user and group names using PMAPI calls
This tackles a failure in CI of test qa/1978 due to different
user and group names being reported on test systems.  The root
cause turned out to be use of getpwuid and getgrgid inside the
atop printing routines.  These map IDs to localhost names - we
can do better and use the names from the archive or (possibly
remote) pmcd, which contains the actual map from the collector
host.