debian packaging: cull old conffiles
The Debian packaging policy dictates that if a conffile,
e.g. /etc/pcp/foo.conf  is included in some verison of an installed
package, and when the package is updated if conffile is no longer in
the package, then the file is not removed and the association between
the package and the file is retained in the dpkg database.

The rationale here is a bit obtuse, but there is a reasonable
explanation at https://wiki.debian.org/DpkgConffileHandling

Note this applies only to "conffiles", but with current versions of the
Debian packaging tools that means pretty much everything below /etc!

There is a documented, but tortuous, path for ensuring that conffiles
that are really retired are removed on package update, but we've
not been implementing this in PCP land.  So we have a long history
of files that remain in the filesystem long after they've vanished
from PCP packages.

For the most part this is benign, but recent changes for the PCP
systemd integration have exposed this as a  problem.

This commit tries to repair the historical damage, and provides a
platform for doing it the "right way" moving forward.

Also included is a new qa/1346 to check for signs of badness in
this area.